avoid the breoken iwyu warning about stdarg.h
move strlist logging to log.c
