+++ src/util/log.h
@@ -1,7 +1,4 @@
 #pragma once
-#ifdef __cplusplus
-extern "C" {
-#endif
 #include <stdio.h>
 #include <stdarg.h>
 
@@ -15,26 +12,27 @@ extern "C" {
 
 #define ERRNO_CHECK() errno_check(__FILE__, __LINE__)
 
-    struct log_t;
+struct log_t;
+struct strlist;
 
-    typedef void(*log_fun)(void *data, int level, const char *module, const char *format, va_list args);
+typedef void(*log_fun)(void *data, int level, const char *module, const char *format, va_list args);
 
-    struct log_t * log_open(const char *filename, int flags);
-    struct log_t * log_create(int flags, void *data, log_fun call);
-    void log_destroy(struct log_t *handle);
-    struct log_t * log_to_file(int flags, FILE *out);
-    int log_level(struct log_t *log, int flags);
-    void log_close(void);
+extern int log_stderr;
 
-    void log_fatal(const char *format, ...);
-    void log_error(const char *format, ...);
-    void log_warning(const char *format, ...);
-    void log_debug(const char *format, ...);
-    void log_info(const char *format, ...);
+struct log_t * log_open(const char *filename, int flags);
+struct log_t * log_create(int flags, void *data, log_fun call);
+void log_destroy(struct log_t *handle);
+struct log_t * log_to_file(int flags, FILE *out);
+int log_level(struct log_t *log, int flags);
+void log_close(void);
 
-    void errno_check(const char *file, int line);
+void log_fatal(const char *format, ...);
+void log_error(const char *format, ...);
+void log_warning(const char *format, ...);
+void log_debug(const char *format, ...);
+void log_info(const char *format, ...);
 
-    extern int log_stderr;
-#ifdef __cplusplus
-}
-#endif
+void errno_check(const char *file, int line);
+
+struct log_t * test_log_start(int flags, struct strlist **slist);
+void test_log_stop(struct log_t *log, struct strlist *slist);
